builder pattern, toEntity 적용

+++ src/main/java/com/mac/demo/repository/AttachRepository.java
@@ -1,16 +1,15 @@
 package com.mac.demo.repository;
 
-import java.util.List;
-
+import com.mac.demo.dto.Attach;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.mac.demo.dto.Attach;
+import java.util.List;
 
 @Repository
-public interface AttachRepository extends JpaRepository<Attach, Integer> {
+public interface AttachRepository extends JpaRepository<Attach, Long> {
 
-	List<Attach> findAllByPcodemac(int pcodeMac);
+	List<Attach> findAllByPcodemac(Long pcodeMac);
 	
 
 }
